Fix : 빌드 에러 수정

+++ src/main/java/com/example/demo/src/record/model/PostRecordRes.java
@@ -7,8 +7,8 @@ import lombok.*;
 @AllArgsConstructor // 해당 클래스의 모든 멤버 변수(email, password)를 받는 생성자를 생성
 @NoArgsConstructor(access = AccessLevel.PROTECTED)  // 해당 클래스의 파라미터가 없는 생성자를 생성, 접근제한자를 PROTECTED로 설정.
 public class PostRecordRes {
-    int userIdx;
-    int goalIdx;
+    Long userIdx;
+    Long goalIdx;
     String date;
-    int recordIdx;
+    Long recordIdx;
 }
